requirements.txt
pandas>=2.0
reportlab>=4.0
PyPDF2>=3.0
openpyxl>=3.1
xlrd>=2.0
qrcode[pil]>=7.4
Pillow>=10.0
flask>=3.0
pyyaml>=6.0
